feat(applicative): add stub for Applicative class interface
Since Typescript doesn't allow static member interfaces, Applicative1 can't really enforce an
instance, but at the very least the comment may be a good indicator to developers of what should be
done with it. Monad now extends it. This should also be useful for creating abstract functions which
create or return Applicatives.

relates #123